gcp-bucket-transfer/requirements.txt
google-cloud-storage==1.9.0
dictdiffer==0.7.1
pytz==2018.5
